Correction for meta-analysis with multiple biases

Correction for meta-analysis with multiple biases

Usage

multibias_meta(
  yi,
  vi,
  sei,
  cluster = 1:length(yi),
  biased = TRUE,
  selection_ratio,
  bias_affirmative,
  bias_nonaffirmative,
  favor_positive = TRUE,
  alpha_select = 0.05,
  ci_level = 0.95,
  small = TRUE,
  return_worst_meta = FALSE,
  return_pubbias_meta = FALSE
)

Arguments

yi

A vector of point estimates to be meta-analyzed.

vi

A vector of estimated variances (i.e., squared standard errors) for the point estimates.

sei

A vector of estimated standard errors for the point estimates. (Only one of vi or sei needs to be specified).

cluster

Vector of the same length as the number of rows in the data, indicating which cluster each study should be considered part of (defaults to treating studies as independent; i.e., each study is in its own cluster).

biased

Boolean indicating whether each study is considered internally biased; either single value used for all studies or a vector the same length as the number of rows in the data (defaults to all studies).

selection_ratio

Ratio by which publication bias favors affirmative studies (i.e., studies with p-values less than alpha_select and estimates in the direction indicated by favor_positive).

bias_affirmative

Mean internal bias, on the additive scale, among published affirmative studies. The bias has the same units as yi.

bias_nonaffirmative

Mean internal bias, on the additive scale, among published nonaffirmative studies. The bias has the same units as yi.

favor_positive

TRUE if publication bias are assumed to favor significant positive estimates; FALSE if assumed to favor significant negative estimates.

alpha_select

Alpha level at which an estimate's probability of being favored by publication bias is assumed to change (i.e., the threshold at which study investigators, journal editors, etc., consider an estimate to be significant).

ci_level

Confidence interval level (as proportion) for the corrected point estimate. (The alpha level for inference on the corrected point estimate will be calculated from ci_level.)

small

Should inference allow for a small meta-analysis? We recommend always using TRUE.

return_worst_meta

Boolean indicating whether the worst-case meta-analysis of only the nonaffirmative studies be returned.

return_pubbias_meta

Boolean indicating whether a meta-analysis correcting for publication but not for confounding be returned.

Value

An object of class metabias::metabias(), a list containing:

data

A tibble with one row per study and the columns yi, vi, sei, biased, cluster, affirmative, yi_adj, weight, userweight.

values

A list with the elements selection_ratio, bias_affirmative, bias_nonaffirmative, favor_positive, alpha_select, ci_level, small.

stats

A tibble with the columns model, estimate, se, ci_lower, ci_upper, p_value.

fit

A list of fitted models.

References

Mathur MB (2022). “Sensitivity analysis for the interactive effects of internal bias and publication bias in meta-analyses.” doi:10.31219/osf.io/u7vcb , osf.io/u7vcb.

Examples

# publication bias without internal bias
multibias_meta(yi = meta_meat$yi,
               vi = meta_meat$vi,
               biased = !meta_meat$randomized,
               selection_ratio = 4,
               bias_affirmative = 0,
               bias_nonaffirmative = 0)
#> List of 4
#>  $ data  : tibble [100 × 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ yi         : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   ..$ vi         : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   ..$ sei        : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   ..$ biased     : logi [1:100] TRUE TRUE TRUE FALSE FALSE FALSE ...
#>   ..$ cluster    : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ affirmative: logi [1:100] FALSE FALSE FALSE FALSE FALSE TRUE ...
#>   ..$ yi_adj     : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   ..$ weight     : num [1:100] 4 4 4 4 4 1 1 4 1 4 ...
#>   ..$ userweight : num [1:100] 26 83.5 72.2 41.5 87.5 ...
#>  $ values:List of 7
#>   ..$ selection_ratio    : num 4
#>   ..$ bias_affirmative   : num 0
#>   ..$ bias_nonaffirmative: num 0
#>   ..$ favor_positive     : logi TRUE
#>   ..$ alpha_select       : num 0.05
#>   ..$ ci_level           : num 0.95
#>   ..$ small              : logi TRUE
#>  $ stats :'data.frame':	1 obs. of  6 variables:
#>   ..$ model   : chr "multibias"
#>   ..$ estimate: num 0.129
#>   ..$ se      : num 0.021
#>   ..$ ci_lower: num 0.0871
#>   ..$ ci_upper: num 0.172
#>   ..$ p_value : num 1.34e-07
#>  $ fits  :List of 1
#>   ..$ multibias:List of 25
#>   .. ..$ data.full     :'data.frame':	100 obs. of  14 variables:
#>   .. .. ..$ effect.size     : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   .. .. ..$ var.eff.size    : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ userweights     : num [1:100] 26 83.5 72.2 41.5 87.5 ...
#>   .. .. ..$ study           : num [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ k               : int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ avg.var.eff.size: num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ sd.eff.size     : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   .. .. ..$ weights         : num [1:100] 8.78 122.29 63.99 17.64 166.67 ...
#>   .. .. ..$ pred            : num [1:100, 1] 0.2 0.2 0.2 0.2 0.2 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ e               : num [1:100, 1] -0.5887 -0.199 -0.2783 0.1947 -0.0912 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ r.weights       : num [1:100] 26 83.5 72.2 41.5 87.5 ...
#>   .. .. ..$ pred.r          : num [1:100, 1] 0.129 0.129 0.129 0.129 0.129 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ e.r             : num [1:100] -0.5177 -0.1279 -0.2072 0.2658 -0.0202 ...
#>   .. .. ..$ ImH             : num [1:100, 1:100] 0.99367 -0.00633 -0.00633 -0.00633 -0.00633 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. ..$ X.full        :'data.frame':	100 obs. of  2 variables:
#>   .. .. ..$ study       : num [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ X.Intercept.: num [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ reg_table     :'data.frame':	1 obs. of  9 variables:
#>   .. .. ..$ labels: chr "X.Intercept."
#>   .. .. ..$ b.r   : num 0.129
#>   .. .. ..$ SE    : num 0.021
#>   .. .. ..$ t     : num 6.15
#>   .. .. ..$ dfs   : num 49.3
#>   .. .. ..$ prob  : num 1.34e-07
#>   .. .. ..$ CI.L  : num 0.0871
#>   .. .. ..$ CI.U  : num 0.172
#>   .. .. ..$ sig   : chr "***"
#>   .. ..$ mod_label     : chr [1:2] "RVE: User Specified Weights" "with Small-Sample Corrections"
#>   .. ..$ mod_notice    : chr "Note: If df < 4, do not trust the results"
#>   .. ..$ modelweights  : chr "CORR"
#>   .. ..$ mod_info      :List of 4
#>   .. .. ..$ k: int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ N: num 100
#>   .. .. ..$ p: num 0
#>   .. .. ..$ M: int 100
#>   .. ..$ user_weighting: logi TRUE
#>   .. ..$ ml            : language yi_adj ~ 1
#>   .. ..$ cl            : language robumeta::robu(formula = yi_adj ~ 1, data = dat, studynum = cluster, var.eff.size = dat$vi,      userweights = da| __truncated__
#>   .. ..$ N             : num 100
#>   .. ..$ M             : int 100
#>   .. ..$ k             : int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ k_list        :List of 100
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. .. [list output truncated]
#>   .. ..$ p             : num 0
#>   .. ..$ X             :List of 100
#>   .. .. ..$ 1  : num [1, 1] 1
#>   .. .. ..$ 2  : num [1, 1] 1
#>   .. .. ..$ 3  : num [1, 1] 1
#>   .. .. ..$ 4  : num [1, 1] 1
#>   .. .. ..$ 5  : num [1, 1] 1
#>   .. .. ..$ 6  : num [1, 1] 1
#>   .. .. ..$ 7  : num [1, 1] 1
#>   .. .. ..$ 8  : num [1, 1] 1
#>   .. .. ..$ 9  : num [1, 1] 1
#>   .. .. ..$ 10 : num [1, 1] 1
#>   .. .. ..$ 11 : num [1, 1] 1
#>   .. .. ..$ 12 : num [1, 1] 1
#>   .. .. ..$ 13 : num [1, 1] 1
#>   .. .. ..$ 14 : num [1, 1] 1
#>   .. .. ..$ 15 : num [1, 1] 1
#>   .. .. ..$ 16 : num [1, 1] 1
#>   .. .. ..$ 17 : num [1, 1] 1
#>   .. .. ..$ 18 : num [1, 1] 1
#>   .. .. ..$ 19 : num [1, 1] 1
#>   .. .. ..$ 20 : num [1, 1] 1
#>   .. .. ..$ 21 : num [1, 1] 1
#>   .. .. ..$ 22 : num [1, 1] 1
#>   .. .. ..$ 23 : num [1, 1] 1
#>   .. .. ..$ 24 : num [1, 1] 1
#>   .. .. ..$ 25 : num [1, 1] 1
#>   .. .. ..$ 26 : num [1, 1] 1
#>   .. .. ..$ 27 : num [1, 1] 1
#>   .. .. ..$ 28 : num [1, 1] 1
#>   .. .. ..$ 29 : num [1, 1] 1
#>   .. .. ..$ 30 : num [1, 1] 1
#>   .. .. ..$ 31 : num [1, 1] 1
#>   .. .. ..$ 32 : num [1, 1] 1
#>   .. .. ..$ 33 : num [1, 1] 1
#>   .. .. ..$ 34 : num [1, 1] 1
#>   .. .. ..$ 35 : num [1, 1] 1
#>   .. .. ..$ 36 : num [1, 1] 1
#>   .. .. ..$ 37 : num [1, 1] 1
#>   .. .. ..$ 38 : num [1, 1] 1
#>   .. .. ..$ 39 : num [1, 1] 1
#>   .. .. ..$ 40 : num [1, 1] 1
#>   .. .. ..$ 41 : num [1, 1] 1
#>   .. .. ..$ 42 : num [1, 1] 1
#>   .. .. ..$ 43 : num [1, 1] 1
#>   .. .. ..$ 44 : num [1, 1] 1
#>   .. .. ..$ 45 : num [1, 1] 1
#>   .. .. ..$ 46 : num [1, 1] 1
#>   .. .. ..$ 47 : num [1, 1] 1
#>   .. .. ..$ 48 : num [1, 1] 1
#>   .. .. ..$ 49 : num [1, 1] 1
#>   .. .. ..$ 50 : num [1, 1] 1
#>   .. .. ..$ 51 : num [1, 1] 1
#>   .. .. ..$ 52 : num [1, 1] 1
#>   .. .. ..$ 53 : num [1, 1] 1
#>   .. .. ..$ 54 : num [1, 1] 1
#>   .. .. ..$ 55 : num [1, 1] 1
#>   .. .. ..$ 56 : num [1, 1] 1
#>   .. .. ..$ 57 : num [1, 1] 1
#>   .. .. ..$ 58 : num [1, 1] 1
#>   .. .. ..$ 59 : num [1, 1] 1
#>   .. .. ..$ 60 : num [1, 1] 1
#>   .. .. ..$ 61 : num [1, 1] 1
#>   .. .. ..$ 62 : num [1, 1] 1
#>   .. .. ..$ 63 : num [1, 1] 1
#>   .. .. ..$ 64 : num [1, 1] 1
#>   .. .. ..$ 65 : num [1, 1] 1
#>   .. .. ..$ 66 : num [1, 1] 1
#>   .. .. ..$ 67 : num [1, 1] 1
#>   .. .. ..$ 68 : num [1, 1] 1
#>   .. .. ..$ 69 : num [1, 1] 1
#>   .. .. ..$ 70 : num [1, 1] 1
#>   .. .. ..$ 71 : num [1, 1] 1
#>   .. .. ..$ 72 : num [1, 1] 1
#>   .. .. ..$ 73 : num [1, 1] 1
#>   .. .. ..$ 74 : num [1, 1] 1
#>   .. .. ..$ 75 : num [1, 1] 1
#>   .. .. ..$ 76 : num [1, 1] 1
#>   .. .. ..$ 77 : num [1, 1] 1
#>   .. .. ..$ 78 : num [1, 1] 1
#>   .. .. ..$ 79 : num [1, 1] 1
#>   .. .. ..$ 80 : num [1, 1] 1
#>   .. .. ..$ 81 : num [1, 1] 1
#>   .. .. ..$ 82 : num [1, 1] 1
#>   .. .. ..$ 83 : num [1, 1] 1
#>   .. .. ..$ 84 : num [1, 1] 1
#>   .. .. ..$ 85 : num [1, 1] 1
#>   .. .. ..$ 86 : num [1, 1] 1
#>   .. .. ..$ 87 : num [1, 1] 1
#>   .. .. ..$ 88 : num [1, 1] 1
#>   .. .. ..$ 89 : num [1, 1] 1
#>   .. .. ..$ 90 : num [1, 1] 1
#>   .. .. ..$ 91 : num [1, 1] 1
#>   .. .. ..$ 92 : num [1, 1] 1
#>   .. .. ..$ 93 : num [1, 1] 1
#>   .. .. ..$ 94 : num [1, 1] 1
#>   .. .. ..$ 95 : num [1, 1] 1
#>   .. .. ..$ 96 : num [1, 1] 1
#>   .. .. ..$ 97 : num [1, 1] 1
#>   .. .. ..$ 98 : num [1, 1] 1
#>   .. .. ..$ 99 : num [1, 1] 1
#>   .. .. .. [list output truncated]
#>   .. ..$ y             : 'by' num [1:100(1d)] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   .. .. ..- attr(*, "dimnames")=List of 1
#>   .. .. .. ..$ data.full$study: chr [1:100] "1" "2" "3" "4" ...
#>   .. .. ..- attr(*, "call")= language by.default(data = data.full$effect.size, INDICES = data.full$study, FUN = function(x) matrix(x))
#>   .. ..$ Xreg          : num [1:100, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. ..$ : chr "X.Intercept."
#>   .. ..$ b.r           : num [1, 1] 0.129
#>   .. ..$ VR.r          : num [1, 1] 0.000442
#>   .. ..$ dfs           : num 49.3
#>   .. ..$ small         : logi TRUE
#>   .. ..$ data          : tibble [100 × 9] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ yi         : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   .. .. ..$ vi         : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ sei        : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   .. .. ..$ biased     : logi [1:100] TRUE TRUE TRUE FALSE FALSE FALSE ...
#>   .. .. ..$ cluster    : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ affirmative: logi [1:100] FALSE FALSE FALSE FALSE FALSE TRUE ...
#>   .. .. ..$ yi_adj     : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   .. .. ..$ weight     : num [1:100] 4 4 4 4 4 1 1 4 1 4 ...
#>   .. .. ..$ userweight : num [1:100] 26 83.5 72.2 41.5 87.5 ...
#>   .. ..$ labels        : chr "X.Intercept."
#>   .. ..$ study_orig_id : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..- attr(*, "class")= chr "robu"
#>  - attr(*, "class")= chr [1:2] "metabias" "list"

# publication bias and internal bias in the non-randomized studies
multibias_meta(yi = meta_meat$yi,
               vi = meta_meat$vi,
               biased = !meta_meat$randomized,
               selection_ratio = 4,
               bias_affirmative = log(1.5),
               bias_nonaffirmative = log(1.1))
#> List of 4
#>  $ data  : tibble [100 × 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ yi         : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   ..$ vi         : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   ..$ sei        : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   ..$ biased     : logi [1:100] TRUE TRUE TRUE FALSE FALSE FALSE ...
#>   ..$ cluster    : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ affirmative: logi [1:100] FALSE FALSE FALSE FALSE FALSE TRUE ...
#>   ..$ yi_adj     : num [1:100] -0.514 -0.124 -0.204 0.395 0.109 ...
#>   ..$ weight     : num [1:100] 4 4 4 4 4 1 1 4 1 4 ...
#>   ..$ userweight : num [1:100] 25.4 77.1 67.4 39.8 80.5 ...
#>  $ values:List of 7
#>   ..$ selection_ratio    : num 4
#>   ..$ bias_affirmative   : num 0.405
#>   ..$ bias_nonaffirmative: num 0.0953
#>   ..$ favor_positive     : logi TRUE
#>   ..$ alpha_select       : num 0.05
#>   ..$ ci_level           : num 0.95
#>   ..$ small              : logi TRUE
#>  $ stats :'data.frame':	1 obs. of  6 variables:
#>   ..$ model   : chr "multibias"
#>   ..$ estimate: num 0.0949
#>   ..$ se      : num 0.0241
#>   ..$ ci_lower: num 0.0466
#>   ..$ ci_upper: num 0.143
#>   ..$ p_value : num 0.000245
#>  $ fits  :List of 1
#>   ..$ multibias:List of 25
#>   .. ..$ data.full     :'data.frame':	100 obs. of  14 variables:
#>   .. .. ..$ effect.size     : num [1:100] -0.514 -0.124 -0.204 0.395 0.109 ...
#>   .. .. ..$ var.eff.size    : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ userweights     : num [1:100] 25.4 77.1 67.4 39.8 80.5 ...
#>   .. .. ..$ study           : num [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ k               : int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ avg.var.eff.size: num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ sd.eff.size     : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   .. .. ..$ weights         : num [1:100] 8.78 122.29 63.99 17.64 166.67 ...
#>   .. .. ..$ pred            : num [1:100, 1] 0.156 0.156 0.156 0.156 0.156 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ e               : num [1:100, 1] -0.6698 -0.28 -0.3593 0.2396 -0.0464 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ r.weights       : num [1:100] 25.4 77.1 67.4 39.8 80.5 ...
#>   .. .. ..$ pred.r          : num [1:100, 1] 0.0949 0.0949 0.0949 0.0949 0.0949 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ e.r             : num [1:100] -0.6091 -0.2194 -0.2986 0.3002 0.0143 ...
#>   .. .. ..$ ImH             : num [1:100, 1:100] 0.99345 -0.00655 -0.00655 -0.00655 -0.00655 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. ..$ X.full        :'data.frame':	100 obs. of  2 variables:
#>   .. .. ..$ study       : num [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ X.Intercept.: num [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ reg_table     :'data.frame':	1 obs. of  9 variables:
#>   .. .. ..$ labels: chr "X.Intercept."
#>   .. .. ..$ b.r   : num 0.0949
#>   .. .. ..$ SE    : num 0.0241
#>   .. .. ..$ t     : num 3.95
#>   .. .. ..$ dfs   : num 50.6
#>   .. .. ..$ prob  : num 0.000245
#>   .. .. ..$ CI.L  : num 0.0466
#>   .. .. ..$ CI.U  : num 0.143
#>   .. .. ..$ sig   : chr "***"
#>   .. ..$ mod_label     : chr [1:2] "RVE: User Specified Weights" "with Small-Sample Corrections"
#>   .. ..$ mod_notice    : chr "Note: If df < 4, do not trust the results"
#>   .. ..$ modelweights  : chr "CORR"
#>   .. ..$ mod_info      :List of 4
#>   .. .. ..$ k: int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ N: num 100
#>   .. .. ..$ p: num 0
#>   .. .. ..$ M: int 100
#>   .. ..$ user_weighting: logi TRUE
#>   .. ..$ ml            : language yi_adj ~ 1
#>   .. ..$ cl            : language robumeta::robu(formula = yi_adj ~ 1, data = dat, studynum = cluster, var.eff.size = dat$vi,      userweights = da| __truncated__
#>   .. ..$ N             : num 100
#>   .. ..$ M             : int 100
#>   .. ..$ k             : int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ k_list        :List of 100
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. .. [list output truncated]
#>   .. ..$ p             : num 0
#>   .. ..$ X             :List of 100
#>   .. .. ..$ 1  : num [1, 1] 1
#>   .. .. ..$ 2  : num [1, 1] 1
#>   .. .. ..$ 3  : num [1, 1] 1
#>   .. .. ..$ 4  : num [1, 1] 1
#>   .. .. ..$ 5  : num [1, 1] 1
#>   .. .. ..$ 6  : num [1, 1] 1
#>   .. .. ..$ 7  : num [1, 1] 1
#>   .. .. ..$ 8  : num [1, 1] 1
#>   .. .. ..$ 9  : num [1, 1] 1
#>   .. .. ..$ 10 : num [1, 1] 1
#>   .. .. ..$ 11 : num [1, 1] 1
#>   .. .. ..$ 12 : num [1, 1] 1
#>   .. .. ..$ 13 : num [1, 1] 1
#>   .. .. ..$ 14 : num [1, 1] 1
#>   .. .. ..$ 15 : num [1, 1] 1
#>   .. .. ..$ 16 : num [1, 1] 1
#>   .. .. ..$ 17 : num [1, 1] 1
#>   .. .. ..$ 18 : num [1, 1] 1
#>   .. .. ..$ 19 : num [1, 1] 1
#>   .. .. ..$ 20 : num [1, 1] 1
#>   .. .. ..$ 21 : num [1, 1] 1
#>   .. .. ..$ 22 : num [1, 1] 1
#>   .. .. ..$ 23 : num [1, 1] 1
#>   .. .. ..$ 24 : num [1, 1] 1
#>   .. .. ..$ 25 : num [1, 1] 1
#>   .. .. ..$ 26 : num [1, 1] 1
#>   .. .. ..$ 27 : num [1, 1] 1
#>   .. .. ..$ 28 : num [1, 1] 1
#>   .. .. ..$ 29 : num [1, 1] 1
#>   .. .. ..$ 30 : num [1, 1] 1
#>   .. .. ..$ 31 : num [1, 1] 1
#>   .. .. ..$ 32 : num [1, 1] 1
#>   .. .. ..$ 33 : num [1, 1] 1
#>   .. .. ..$ 34 : num [1, 1] 1
#>   .. .. ..$ 35 : num [1, 1] 1
#>   .. .. ..$ 36 : num [1, 1] 1
#>   .. .. ..$ 37 : num [1, 1] 1
#>   .. .. ..$ 38 : num [1, 1] 1
#>   .. .. ..$ 39 : num [1, 1] 1
#>   .. .. ..$ 40 : num [1, 1] 1
#>   .. .. ..$ 41 : num [1, 1] 1
#>   .. .. ..$ 42 : num [1, 1] 1
#>   .. .. ..$ 43 : num [1, 1] 1
#>   .. .. ..$ 44 : num [1, 1] 1
#>   .. .. ..$ 45 : num [1, 1] 1
#>   .. .. ..$ 46 : num [1, 1] 1
#>   .. .. ..$ 47 : num [1, 1] 1
#>   .. .. ..$ 48 : num [1, 1] 1
#>   .. .. ..$ 49 : num [1, 1] 1
#>   .. .. ..$ 50 : num [1, 1] 1
#>   .. .. ..$ 51 : num [1, 1] 1
#>   .. .. ..$ 52 : num [1, 1] 1
#>   .. .. ..$ 53 : num [1, 1] 1
#>   .. .. ..$ 54 : num [1, 1] 1
#>   .. .. ..$ 55 : num [1, 1] 1
#>   .. .. ..$ 56 : num [1, 1] 1
#>   .. .. ..$ 57 : num [1, 1] 1
#>   .. .. ..$ 58 : num [1, 1] 1
#>   .. .. ..$ 59 : num [1, 1] 1
#>   .. .. ..$ 60 : num [1, 1] 1
#>   .. .. ..$ 61 : num [1, 1] 1
#>   .. .. ..$ 62 : num [1, 1] 1
#>   .. .. ..$ 63 : num [1, 1] 1
#>   .. .. ..$ 64 : num [1, 1] 1
#>   .. .. ..$ 65 : num [1, 1] 1
#>   .. .. ..$ 66 : num [1, 1] 1
#>   .. .. ..$ 67 : num [1, 1] 1
#>   .. .. ..$ 68 : num [1, 1] 1
#>   .. .. ..$ 69 : num [1, 1] 1
#>   .. .. ..$ 70 : num [1, 1] 1
#>   .. .. ..$ 71 : num [1, 1] 1
#>   .. .. ..$ 72 : num [1, 1] 1
#>   .. .. ..$ 73 : num [1, 1] 1
#>   .. .. ..$ 74 : num [1, 1] 1
#>   .. .. ..$ 75 : num [1, 1] 1
#>   .. .. ..$ 76 : num [1, 1] 1
#>   .. .. ..$ 77 : num [1, 1] 1
#>   .. .. ..$ 78 : num [1, 1] 1
#>   .. .. ..$ 79 : num [1, 1] 1
#>   .. .. ..$ 80 : num [1, 1] 1
#>   .. .. ..$ 81 : num [1, 1] 1
#>   .. .. ..$ 82 : num [1, 1] 1
#>   .. .. ..$ 83 : num [1, 1] 1
#>   .. .. ..$ 84 : num [1, 1] 1
#>   .. .. ..$ 85 : num [1, 1] 1
#>   .. .. ..$ 86 : num [1, 1] 1
#>   .. .. ..$ 87 : num [1, 1] 1
#>   .. .. ..$ 88 : num [1, 1] 1
#>   .. .. ..$ 89 : num [1, 1] 1
#>   .. .. ..$ 90 : num [1, 1] 1
#>   .. .. ..$ 91 : num [1, 1] 1
#>   .. .. ..$ 92 : num [1, 1] 1
#>   .. .. ..$ 93 : num [1, 1] 1
#>   .. .. ..$ 94 : num [1, 1] 1
#>   .. .. ..$ 95 : num [1, 1] 1
#>   .. .. ..$ 96 : num [1, 1] 1
#>   .. .. ..$ 97 : num [1, 1] 1
#>   .. .. ..$ 98 : num [1, 1] 1
#>   .. .. ..$ 99 : num [1, 1] 1
#>   .. .. .. [list output truncated]
#>   .. ..$ y             : 'by' num [1:100(1d)] -0.514 -0.124 -0.204 0.395 0.109 ...
#>   .. .. ..- attr(*, "dimnames")=List of 1
#>   .. .. .. ..$ data.full$study: chr [1:100] "1" "2" "3" "4" ...
#>   .. .. ..- attr(*, "call")= language by.default(data = data.full$effect.size, INDICES = data.full$study, FUN = function(x) matrix(x))
#>   .. ..$ Xreg          : num [1:100, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. ..$ : chr "X.Intercept."
#>   .. ..$ b.r           : num [1, 1] 0.0949
#>   .. ..$ VR.r          : num [1, 1] 0.000578
#>   .. ..$ dfs           : num 50.6
#>   .. ..$ small         : logi TRUE
#>   .. ..$ data          : tibble [100 × 9] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ yi         : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   .. .. ..$ vi         : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ sei        : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   .. .. ..$ biased     : logi [1:100] TRUE TRUE TRUE FALSE FALSE FALSE ...
#>   .. .. ..$ cluster    : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ affirmative: logi [1:100] FALSE FALSE FALSE FALSE FALSE TRUE ...
#>   .. .. ..$ yi_adj     : num [1:100] -0.514 -0.124 -0.204 0.395 0.109 ...
#>   .. .. ..$ weight     : num [1:100] 4 4 4 4 4 1 1 4 1 4 ...
#>   .. .. ..$ userweight : num [1:100] 25.4 77.1 67.4 39.8 80.5 ...
#>   .. ..$ labels        : chr "X.Intercept."
#>   .. ..$ study_orig_id : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..- attr(*, "class")= chr "robu"
#>  - attr(*, "class")= chr [1:2] "metabias" "list"

# treat all studies as biased, not just non-randomized ones
multibias_meta(yi = meta_meat$yi,
               vi = meta_meat$vi,
               biased = TRUE,
               selection_ratio = 4,
               bias_affirmative = log(1.5),
               bias_nonaffirmative = log(1.1))
#> List of 4
#>  $ data  : tibble [100 × 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ yi         : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   ..$ vi         : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   ..$ sei        : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   ..$ biased     : logi [1:100] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ cluster    : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ affirmative: logi [1:100] FALSE FALSE FALSE FALSE FALSE TRUE ...
#>   ..$ yi_adj     : num [1:100] -0.50747 -0.11774 -0.19699 0.27596 -0.00997 ...
#>   ..$ weight     : num [1:100] 4 4 4 4 4 1 1 4 1 4 ...
#>   ..$ userweight : num [1:100] 26 83.5 72.2 41.5 87.5 ...
#>  $ values:List of 7
#>   ..$ selection_ratio    : num 4
#>   ..$ bias_affirmative   : num 0.405
#>   ..$ bias_nonaffirmative: num 0.0953
#>   ..$ favor_positive     : logi TRUE
#>   ..$ alpha_select       : num 0.05
#>   ..$ ci_level           : num 0.95
#>   ..$ small              : logi TRUE
#>  $ stats :'data.frame':	1 obs. of  6 variables:
#>   ..$ model   : chr "multibias"
#>   ..$ estimate: num 0.0102
#>   ..$ se      : num 0.021
#>   ..$ ci_lower: num -0.0321
#>   ..$ ci_upper: num 0.0525
#>   ..$ p_value : num 0.63
#>  $ fits  :List of 1
#>   ..$ multibias:List of 25
#>   .. ..$ data.full     :'data.frame':	100 obs. of  14 variables:
#>   .. .. ..$ effect.size     : num [1:100] -0.50747 -0.11774 -0.19699 0.27596 -0.00997 ...
#>   .. .. ..$ var.eff.size    : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ userweights     : num [1:100] 26 83.5 72.2 41.5 87.5 ...
#>   .. .. ..$ study           : num [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ k               : int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ avg.var.eff.size: num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ sd.eff.size     : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   .. .. ..$ weights         : num [1:100] 8.78 122.29 63.99 17.64 166.67 ...
#>   .. .. ..$ pred            : num [1:100, 1] 0.0813 0.0813 0.0813 0.0813 0.0813 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ e               : num [1:100, 1] -0.5887 -0.199 -0.2783 0.1947 -0.0912 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ r.weights       : num [1:100] 26 83.5 72.2 41.5 87.5 ...
#>   .. .. ..$ pred.r          : num [1:100, 1] 0.0102 0.0102 0.0102 0.0102 0.0102 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. .. ..$ e.r             : num [1:100] -0.5177 -0.1279 -0.2072 0.2658 -0.0202 ...
#>   .. .. ..$ ImH             : num [1:100, 1:100] 0.99367 -0.00633 -0.00633 -0.00633 -0.00633 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. .. ..$ : NULL
#>   .. ..$ X.full        :'data.frame':	100 obs. of  2 variables:
#>   .. .. ..$ study       : num [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ X.Intercept.: num [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ reg_table     :'data.frame':	1 obs. of  9 variables:
#>   .. .. ..$ labels: chr "X.Intercept."
#>   .. .. ..$ b.r   : num 0.0102
#>   .. .. ..$ SE    : num 0.021
#>   .. .. ..$ t     : num 0.485
#>   .. .. ..$ dfs   : num 49.3
#>   .. .. ..$ prob  : num 0.63
#>   .. .. ..$ CI.L  : num -0.0321
#>   .. .. ..$ CI.U  : num 0.0525
#>   .. .. ..$ sig   : chr ""
#>   .. ..$ mod_label     : chr [1:2] "RVE: User Specified Weights" "with Small-Sample Corrections"
#>   .. ..$ mod_notice    : chr "Note: If df < 4, do not trust the results"
#>   .. ..$ modelweights  : chr "CORR"
#>   .. ..$ mod_info      :List of 4
#>   .. .. ..$ k: int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ N: num 100
#>   .. .. ..$ p: num 0
#>   .. .. ..$ M: int 100
#>   .. ..$ user_weighting: logi TRUE
#>   .. ..$ ml            : language yi_adj ~ 1
#>   .. ..$ cl            : language robumeta::robu(formula = yi_adj ~ 1, data = dat, studynum = cluster, var.eff.size = dat$vi,      userweights = da| __truncated__
#>   .. ..$ N             : num 100
#>   .. ..$ M             : int 100
#>   .. ..$ k             : int [1:100] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ k_list        :List of 100
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
#>   .. .. .. [list output truncated]
#>   .. ..$ p             : num 0
#>   .. ..$ X             :List of 100
#>   .. .. ..$ 1  : num [1, 1] 1
#>   .. .. ..$ 2  : num [1, 1] 1
#>   .. .. ..$ 3  : num [1, 1] 1
#>   .. .. ..$ 4  : num [1, 1] 1
#>   .. .. ..$ 5  : num [1, 1] 1
#>   .. .. ..$ 6  : num [1, 1] 1
#>   .. .. ..$ 7  : num [1, 1] 1
#>   .. .. ..$ 8  : num [1, 1] 1
#>   .. .. ..$ 9  : num [1, 1] 1
#>   .. .. ..$ 10 : num [1, 1] 1
#>   .. .. ..$ 11 : num [1, 1] 1
#>   .. .. ..$ 12 : num [1, 1] 1
#>   .. .. ..$ 13 : num [1, 1] 1
#>   .. .. ..$ 14 : num [1, 1] 1
#>   .. .. ..$ 15 : num [1, 1] 1
#>   .. .. ..$ 16 : num [1, 1] 1
#>   .. .. ..$ 17 : num [1, 1] 1
#>   .. .. ..$ 18 : num [1, 1] 1
#>   .. .. ..$ 19 : num [1, 1] 1
#>   .. .. ..$ 20 : num [1, 1] 1
#>   .. .. ..$ 21 : num [1, 1] 1
#>   .. .. ..$ 22 : num [1, 1] 1
#>   .. .. ..$ 23 : num [1, 1] 1
#>   .. .. ..$ 24 : num [1, 1] 1
#>   .. .. ..$ 25 : num [1, 1] 1
#>   .. .. ..$ 26 : num [1, 1] 1
#>   .. .. ..$ 27 : num [1, 1] 1
#>   .. .. ..$ 28 : num [1, 1] 1
#>   .. .. ..$ 29 : num [1, 1] 1
#>   .. .. ..$ 30 : num [1, 1] 1
#>   .. .. ..$ 31 : num [1, 1] 1
#>   .. .. ..$ 32 : num [1, 1] 1
#>   .. .. ..$ 33 : num [1, 1] 1
#>   .. .. ..$ 34 : num [1, 1] 1
#>   .. .. ..$ 35 : num [1, 1] 1
#>   .. .. ..$ 36 : num [1, 1] 1
#>   .. .. ..$ 37 : num [1, 1] 1
#>   .. .. ..$ 38 : num [1, 1] 1
#>   .. .. ..$ 39 : num [1, 1] 1
#>   .. .. ..$ 40 : num [1, 1] 1
#>   .. .. ..$ 41 : num [1, 1] 1
#>   .. .. ..$ 42 : num [1, 1] 1
#>   .. .. ..$ 43 : num [1, 1] 1
#>   .. .. ..$ 44 : num [1, 1] 1
#>   .. .. ..$ 45 : num [1, 1] 1
#>   .. .. ..$ 46 : num [1, 1] 1
#>   .. .. ..$ 47 : num [1, 1] 1
#>   .. .. ..$ 48 : num [1, 1] 1
#>   .. .. ..$ 49 : num [1, 1] 1
#>   .. .. ..$ 50 : num [1, 1] 1
#>   .. .. ..$ 51 : num [1, 1] 1
#>   .. .. ..$ 52 : num [1, 1] 1
#>   .. .. ..$ 53 : num [1, 1] 1
#>   .. .. ..$ 54 : num [1, 1] 1
#>   .. .. ..$ 55 : num [1, 1] 1
#>   .. .. ..$ 56 : num [1, 1] 1
#>   .. .. ..$ 57 : num [1, 1] 1
#>   .. .. ..$ 58 : num [1, 1] 1
#>   .. .. ..$ 59 : num [1, 1] 1
#>   .. .. ..$ 60 : num [1, 1] 1
#>   .. .. ..$ 61 : num [1, 1] 1
#>   .. .. ..$ 62 : num [1, 1] 1
#>   .. .. ..$ 63 : num [1, 1] 1
#>   .. .. ..$ 64 : num [1, 1] 1
#>   .. .. ..$ 65 : num [1, 1] 1
#>   .. .. ..$ 66 : num [1, 1] 1
#>   .. .. ..$ 67 : num [1, 1] 1
#>   .. .. ..$ 68 : num [1, 1] 1
#>   .. .. ..$ 69 : num [1, 1] 1
#>   .. .. ..$ 70 : num [1, 1] 1
#>   .. .. ..$ 71 : num [1, 1] 1
#>   .. .. ..$ 72 : num [1, 1] 1
#>   .. .. ..$ 73 : num [1, 1] 1
#>   .. .. ..$ 74 : num [1, 1] 1
#>   .. .. ..$ 75 : num [1, 1] 1
#>   .. .. ..$ 76 : num [1, 1] 1
#>   .. .. ..$ 77 : num [1, 1] 1
#>   .. .. ..$ 78 : num [1, 1] 1
#>   .. .. ..$ 79 : num [1, 1] 1
#>   .. .. ..$ 80 : num [1, 1] 1
#>   .. .. ..$ 81 : num [1, 1] 1
#>   .. .. ..$ 82 : num [1, 1] 1
#>   .. .. ..$ 83 : num [1, 1] 1
#>   .. .. ..$ 84 : num [1, 1] 1
#>   .. .. ..$ 85 : num [1, 1] 1
#>   .. .. ..$ 86 : num [1, 1] 1
#>   .. .. ..$ 87 : num [1, 1] 1
#>   .. .. ..$ 88 : num [1, 1] 1
#>   .. .. ..$ 89 : num [1, 1] 1
#>   .. .. ..$ 90 : num [1, 1] 1
#>   .. .. ..$ 91 : num [1, 1] 1
#>   .. .. ..$ 92 : num [1, 1] 1
#>   .. .. ..$ 93 : num [1, 1] 1
#>   .. .. ..$ 94 : num [1, 1] 1
#>   .. .. ..$ 95 : num [1, 1] 1
#>   .. .. ..$ 96 : num [1, 1] 1
#>   .. .. ..$ 97 : num [1, 1] 1
#>   .. .. ..$ 98 : num [1, 1] 1
#>   .. .. ..$ 99 : num [1, 1] 1
#>   .. .. .. [list output truncated]
#>   .. ..$ y             : 'by' num [1:100(1d)] -0.50747 -0.11774 -0.19699 0.27596 -0.00997 ...
#>   .. .. ..- attr(*, "dimnames")=List of 1
#>   .. .. .. ..$ data.full$study: chr [1:100] "1" "2" "3" "4" ...
#>   .. .. ..- attr(*, "call")= language by.default(data = data.full$effect.size, INDICES = data.full$study, FUN = function(x) matrix(x))
#>   .. ..$ Xreg          : num [1:100, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:100] "1" "2" "3" "4" ...
#>   .. .. .. ..$ : chr "X.Intercept."
#>   .. ..$ b.r           : num [1, 1] 0.0102
#>   .. ..$ VR.r          : num [1, 1] 0.000442
#>   .. ..$ dfs           : num 49.3
#>   .. ..$ small         : logi TRUE
#>   .. ..$ data          : tibble [100 × 9] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ yi         : num [1:100] -0.3883 0.00143 -0.07782 0.39513 0.1092 ...
#>   .. .. ..$ vi         : num [1:100] 0.11387 0.00818 0.01563 0.0567 0.006 ...
#>   .. .. ..$ sei        : num [1:100] 0.3374 0.0904 0.125 0.2381 0.0775 ...
#>   .. .. ..$ biased     : logi [1:100] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   .. .. ..$ cluster    : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. ..$ affirmative: logi [1:100] FALSE FALSE FALSE FALSE FALSE TRUE ...
#>   .. .. ..$ yi_adj     : num [1:100] -0.50747 -0.11774 -0.19699 0.27596 -0.00997 ...
#>   .. .. ..$ weight     : num [1:100] 4 4 4 4 4 1 1 4 1 4 ...
#>   .. .. ..$ userweight : num [1:100] 26 83.5 72.2 41.5 87.5 ...
#>   .. ..$ labels        : chr "X.Intercept."
#>   .. ..$ study_orig_id : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..- attr(*, "class")= chr "robu"
#>  - attr(*, "class")= chr [1:2] "metabias" "list"