Correction for meta-analysis with multiple biases

Correction for meta-analysis with multiple biases

Usage

multibias_meta(
  yi,
  vi,
  sei,
  cluster = 1:length(yi),
  biased = TRUE,
  selection_ratio,
  bias_affirmative,
  bias_nonaffirmative,
  favor_positive = TRUE,
  alpha_select = 0.05,
  ci_level = 0.95,
  small = TRUE,
  return_worst_meta = FALSE,
  return_pubbias_meta = FALSE
)

Arguments

yi

A vector of point estimates to be meta-analyzed.

vi

A vector of estimated variances (i.e., squared standard errors) for the point estimates.

sei

A vector of estimated standard errors for the point estimates. (Only one of vi or sei needs to be specified).

cluster

Vector of the same length as the number of rows in the data, indicating which cluster each study should be considered part of (defaults to treating studies as independent; i.e., each study is in its own cluster).

biased

Boolean indicating whether each study is considered internally biased; either single value used for all studies or a vector the same length as the number of rows in the data (defaults to all studies).

selection_ratio

Ratio by which publication bias favors affirmative studies (i.e., studies with p-values less than alpha_select and estimates in the direction indicated by favor_positive).

bias_affirmative

Mean internal bias, on the additive scale, among published affirmative studies. The bias has the same units as yi.

bias_nonaffirmative

Mean internal bias, on the additive scale, among published nonaffirmative studies. The bias has the same units as yi.

favor_positive

TRUE if publication bias are assumed to favor significant positive estimates; FALSE if assumed to favor significant negative estimates.

alpha_select

Alpha level at which an estimate's probability of being favored by publication bias is assumed to change (i.e., the threshold at which study investigators, journal editors, etc., consider an estimate to be significant).

ci_level

Confidence interval level (as proportion) for the corrected point estimate. (The alpha level for inference on the corrected point estimate will be calculated from ci_level.)

small

Should inference allow for a small meta-analysis? We recommend always using TRUE.

return_worst_meta

Boolean indicating whether the worst-case meta-analysis of only the nonaffirmative studies be returned.

return_pubbias_meta

Boolean indicating whether a meta-analysis correcting for publication but not for confounding be returned.

Value

An object of class metabias::metabias(), a list containing:

data

A tibble with one row per study and the columns yi, vi, sei, biased, cluster, affirmative, yi_adj, weight, userweight.

values

A list with the elements selection_ratio, bias_affirmative, bias_nonaffirmative, favor_positive, alpha_select, ci_level, small.

stats

A tibble with the columns model, estimate, se, ci_lower, ci_upper, p_value.

fit

A list of fitted models.

References

Mathur MB (2022). “Sensitivity analysis for the interactive effects of internal bias and publication bias in meta-analyses.” doi:10.31219/osf.io/u7vcb , osf.io/u7vcb.

Examples

# publication bias without internal bias
multibias_meta(yi = meta_meat$yi,
               vi = meta_meat$vi,
               selection_ratio = 4,
               bias_affirmative = 0,
               bias_nonaffirmative = 0)
#> $data
#> # A tibble: 100 × 9
#>          yi      vi    sei biased cluster affirmative   yi_adj weight userweight
#>       <dbl>   <dbl>  <dbl> <lgl>    <int> <lgl>          <dbl>  <dbl>      <dbl>
#>  1 -0.388   0.114   0.337  TRUE         1 FALSE       -0.388        4       26.0
#>  2  0.00143 0.00818 0.0904 TRUE         2 FALSE        0.00143      4       83.5
#>  3 -0.0778  0.0156  0.125  TRUE         3 FALSE       -0.0778       4       72.2
#>  4  0.395   0.0567  0.238  TRUE         4 FALSE        0.395        4       41.5
#>  5  0.109   0.006   0.0775 TRUE         5 FALSE        0.109        4       87.5
#>  6  0.0711  0.00110 0.0332 TRUE         6 TRUE         0.0711       1       24.5
#>  7  0.177   0.00629 0.0793 TRUE         7 TRUE         0.177        1       21.7
#>  8  0.123   0.0161  0.127  TRUE         8 FALSE        0.123        4       71.6
#>  9  0.310   0.0169  0.130  TRUE         9 TRUE         0.310        1       17.7
#> 10  0.229   0.0154  0.124  TRUE        10 FALSE        0.229        4       72.5
#> # … with 90 more rows
#> 
#> $values
#> $values$selection_ratio
#> [1] 4
#> 
#> $values$bias_affirmative
#> [1] 0
#> 
#> $values$bias_nonaffirmative
#> [1] 0
#> 
#> $values$favor_positive
#> [1] TRUE
#> 
#> $values$alpha_select
#> [1] 0.05
#> 
#> $values$ci_level
#> [1] 0.95
#> 
#> $values$small
#> [1] TRUE
#> 
#> 
#> $stats
#>       model  estimate         se   ci_lower  ci_upper      p_value
#> 1 multibias 0.1293691 0.02103083 0.08711307 0.1716251 1.335523e-07
#> 
#> $fits
#> $fits$multibias
#> RVE: User Specified Weights with Small-Sample Corrections 
#> 
#> Model: yi_adj ~ 1 
#> 
#> Number of studies = 100 
#> Number of outcomes = 100 (min = 1 , mean = 1 , median = 1 , max = 1 )
#>                Estimate StdErr t-value  dfs     P(|t|>) 95% CI.L 95% CI.U Sig
#> 1 X.Intercept.    0.129  0.021    6.15 49.3 0.000000134   0.0871    0.172 ***
#> ---
#> Signif. codes: < .01 *** < .05 ** < .10 *
#> ---
#> Note: If df < 4, do not trust the results
#> 
#> attr(,"class")
#> [1] "metabias" "list"    

# publication bias and internal bias in the non-randomized studies
multibias_meta(yi = meta_meat$yi,
               vi = meta_meat$vi,
               biased = !meta_meat$randomized,
               selection_ratio = 4,
               bias_affirmative = log(1.5),
               bias_nonaffirmative = log(1.1))
#> $data
#> # A tibble: 100 × 9
#>          yi      vi    sei biased cluster affirmative  yi_adj weight userweight
#>       <dbl>   <dbl>  <dbl> <lgl>    <int> <lgl>         <dbl>  <dbl>      <dbl>
#>  1 -0.388   0.114   0.337  TRUE         1 FALSE       -0.514       4       25.4
#>  2  0.00143 0.00818 0.0904 TRUE         2 FALSE       -0.124       4       77.1
#>  3 -0.0778  0.0156  0.125  TRUE         3 FALSE       -0.204       4       67.4
#>  4  0.395   0.0567  0.238  FALSE        4 FALSE        0.395       4       39.8
#>  5  0.109   0.006   0.0775 FALSE        5 FALSE        0.109       4       80.5
#>  6  0.0711  0.00110 0.0332 FALSE        6 TRUE         0.0711      1       22.3
#>  7  0.177   0.00629 0.0793 FALSE        7 TRUE         0.177       1       20.0
#>  8  0.123   0.0161  0.127  FALSE        8 FALSE        0.123       4       66.9
#>  9  0.310   0.0169  0.130  FALSE        9 TRUE         0.310       1       16.5
#> 10  0.229   0.0154  0.124  FALSE       10 FALSE        0.229       4       67.6
#> # … with 90 more rows
#> 
#> $values
#> $values$selection_ratio
#> [1] 4
#> 
#> $values$bias_affirmative
#> [1] 0.4054651
#> 
#> $values$bias_nonaffirmative
#> [1] 0.09531018
#> 
#> $values$favor_positive
#> [1] TRUE
#> 
#> $values$alpha_select
#> [1] 0.05
#> 
#> $values$ci_level
#> [1] 0.95
#> 
#> $values$small
#> [1] TRUE
#> 
#> 
#> $stats
#>       model   estimate         se   ci_lower  ci_upper      p_value
#> 1 multibias 0.09491905 0.02405073 0.04662505 0.1432131 0.0002453537
#> 
#> $fits
#> $fits$multibias
#> RVE: User Specified Weights with Small-Sample Corrections 
#> 
#> Model: yi_adj ~ 1 
#> 
#> Number of studies = 100 
#> Number of outcomes = 100 (min = 1 , mean = 1 , median = 1 , max = 1 )
#>                Estimate StdErr t-value  dfs  P(|t|>) 95% CI.L 95% CI.U Sig
#> 1 X.Intercept.   0.0949 0.0241    3.95 50.6 0.000245   0.0466    0.143 ***
#> ---
#> Signif. codes: < .01 *** < .05 ** < .10 *
#> ---
#> Note: If df < 4, do not trust the results
#> 
#> attr(,"class")
#> [1] "metabias" "list"    

# treat all studies as biased, not just non-randomized ones
multibias_meta(yi = meta_meat$yi,
               vi = meta_meat$vi,
               biased = TRUE,
               selection_ratio = 4,
               bias_affirmative = log(1.5),
               bias_nonaffirmative = log(1.1))
#> $data
#> # A tibble: 100 × 9
#>          yi      vi    sei biased cluster affirmative   yi_adj weight userweight
#>       <dbl>   <dbl>  <dbl> <lgl>    <int> <lgl>          <dbl>  <dbl>      <dbl>
#>  1 -0.388   0.114   0.337  TRUE         1 FALSE       -0.507        4       26.0
#>  2  0.00143 0.00818 0.0904 TRUE         2 FALSE       -0.118        4       83.5
#>  3 -0.0778  0.0156  0.125  TRUE         3 FALSE       -0.197        4       72.2
#>  4  0.395   0.0567  0.238  TRUE         4 FALSE        0.276        4       41.5
#>  5  0.109   0.006   0.0775 TRUE         5 FALSE       -0.00997      4       87.5
#>  6  0.0711  0.00110 0.0332 TRUE         6 TRUE        -0.0480       1       24.5
#>  7  0.177   0.00629 0.0793 TRUE         7 TRUE         0.0574       1       21.7
#>  8  0.123   0.0161  0.127  TRUE         8 FALSE        0.00428      4       71.6
#>  9  0.310   0.0169  0.130  TRUE         9 TRUE         0.191        1       17.7
#> 10  0.229   0.0154  0.124  TRUE        10 FALSE        0.109        4       72.5
#> # … with 90 more rows
#> 
#> $values
#> $values$selection_ratio
#> [1] 4
#> 
#> $values$bias_affirmative
#> [1] 0.4054651
#> 
#> $values$bias_nonaffirmative
#> [1] 0.09531018
#> 
#> $values$favor_positive
#> [1] TRUE
#> 
#> $values$alpha_select
#> [1] 0.05
#> 
#> $values$ci_level
#> [1] 0.95
#> 
#> $values$small
#> [1] TRUE
#> 
#> 
#> $stats
#>       model   estimate         se    ci_lower  ci_upper   p_value
#> 1 multibias 0.01020086 0.02103083 -0.03205518 0.0524569 0.6297933
#> 
#> $fits
#> $fits$multibias
#> RVE: User Specified Weights with Small-Sample Corrections 
#> 
#> Model: yi_adj ~ 1 
#> 
#> Number of studies = 100 
#> Number of outcomes = 100 (min = 1 , mean = 1 , median = 1 , max = 1 )
#>                Estimate StdErr t-value  dfs P(|t|>) 95% CI.L 95% CI.U Sig
#> 1 X.Intercept.   0.0102  0.021   0.485 49.3    0.63  -0.0321   0.0525    
#> ---
#> Signif. codes: < .01 *** < .05 ** < .10 *
#> ---
#> Note: If df < 4, do not trust the results
#> 
#> attr(,"class")
#> [1] "metabias" "list"